# Improve batch allocation stress test